+++ src/buffer.h
@@ -26,7 +26,7 @@ typedef struct Buffer {
     struct stat st;
 
     // Needed for identifying buffers whose filename is NULL
-    unsigned int id;
+    unsigned long id;
 
     size_t nl;
 
@@ -79,7 +79,7 @@ char *short_filename(const char *absolute) XSTRDUP;
 void update_short_filename_cwd(Buffer *b, const char *cwd);
 void update_short_filename(Buffer *b);
 Buffer *find_buffer(const char *abs_filename);
-Buffer *find_buffer_by_id(unsigned int id);
+Buffer *find_buffer_by_id(unsigned long id);
 Buffer *buffer_new(const Encoding *encoding);
 Buffer *open_empty_buffer(void);
 void free_buffer(Buffer *b);
